+++ nodeconductor/cost_tracking/__init__.py
@@ -13,18 +13,17 @@ Add-on connection:
 
     1. Implement CostTrackingBackend interface.
 
-        class IaaSCostTrackingBackend(CostTrackingBackend):
+        class OpenStackCostTrackingBackend(CostTrackingBackend):
             ...
 
     2. Add application to add-on register.
 
-        CostTrackingRegister.register(self.label, cost_tracking.IaaSCostTrackingBackend)
+        CostTrackingRegister.register(self.label, cost_tracking.OpenStackCostTrackingBackend)
 
 """
 import logging
 from decimal import Decimal
 
-from django.contrib.contenttypes.models import ContentType
 from django.db.models import Prefetch
 
 from nodeconductor.structure import ServiceBackendNotImplemented
@@ -113,6 +112,7 @@ class CostTrackingBackend(object):
         If service-specific price list item is unavailable, default
         price list item is used instead for fetching monthly_rate.
         """
+        from django.contrib.contenttypes.models import ContentType
         from nodeconductor.cost_tracking.models import DefaultPriceListItem, PriceListItem
         resource_content_type = ContentType.objects.get_for_model(resource)