+++ nodeconductor/cost_tracking/tests/test_limit.py
@@ -4,8 +4,7 @@ from rest_framework import test
 
 from nodeconductor.cost_tracking.exceptions import CostLimitExceeded
 from nodeconductor.cost_tracking.models import PriceEstimate
-from nodeconductor.openstack.tests.factories import InstanceFactory, OpenStackServiceProjectLinkFactory
-from nodeconductor.structure.tests.factories import ProjectFactory
+from nodeconductor.structure.tests.factories import ProjectFactory, TestInstanceFactory, TestServiceProjectLinkFactory
 
 
 class TestProjectCostLimit(test.APITransactionTestCase):
@@ -28,7 +27,7 @@ class TestProjectCostLimit(test.APITransactionTestCase):
         return project
 
     def create_resource(self, project, cost):
-        link = OpenStackServiceProjectLinkFactory(project=project)
+        link = TestServiceProjectLinkFactory(project=project)
         with mock.patch('nodeconductor.cost_tracking.handlers.CostTrackingRegister') as register:
             register.get_resource_backend().get_monthly_cost_estimate.return_value = cost
-            InstanceFactory(service_project_link=link)
+            TestInstanceFactory(service_project_link=link)