+++ nodeconductor/monitoring/managers.py
@@ -8,3 +8,7 @@ class ResourceSlaManager(GenericKeyMixin, django_models.Manager):
 
 class ResourceItemManager(GenericKeyMixin, django_models.Manager):
     pass
+
+
+class ResourceSlaStateTransitionManager(GenericKeyMixin, django_models.Manager):
+    pass