+++ nodeconductor/server/urls.py
@@ -12,11 +12,8 @@ from django.views.generic import TemplateView
 from nodeconductor.core import NodeConductorExtension
 from nodeconductor.core.routers import SortedDefaultRouter as DefaultRouter
 from nodeconductor.cost_tracking import urls as cost_tracking_urls
-from nodeconductor.backup import urls as backup_urls
-from nodeconductor.iaas import urls as iaas_urls
 from nodeconductor.logging import urls as logging_urls
 from nodeconductor.monitoring import urls as monitoring_urls
-from nodeconductor.openstack import urls as openstack_urls
 from nodeconductor.quotas import urls as quotas_urls
 from nodeconductor.structure import urls as structure_urls
 from nodeconductor.template import urls as template_urls
@@ -26,12 +23,9 @@ admin.autodiscover()
 permission.autodiscover()
 
 router = DefaultRouter()
-backup_urls.register_in(router)
 cost_tracking_urls.register_in(router)
-iaas_urls.register_in(router)
 logging_urls.register_in(router)
 monitoring_urls.register_in(router)
-openstack_urls.register_in(router)
 quotas_urls.register_in(router)
 structure_urls.register_in(router)
 template_urls.register_in(router)
@@ -53,7 +47,6 @@ urlpatterns += patterns(
     '',
     url(r'^api/', include(router.urls)),
     url(r'^api/', include('nodeconductor.logging.urls')),
-    url(r'^api/', include('nodeconductor.iaas.urls')),
     url(r'^api/', include('nodeconductor.structure.urls')),
     url(r'^api/version/', 'nodeconductor.core.views.version_detail'),
     url(r'^api-auth/password/', 'nodeconductor.core.views.obtain_auth_token', name='auth-password'),