+++ nodeconductor/structure/tests/test_service.py
@@ -1,14 +1,9 @@
-import factory
-
 from mock_django import mock_signal_receiver
 from django.core.urlresolvers import reverse
-from rest_framework import test, status
+from rest_framework import test
 
-from nodeconductor.iaas.models import OpenStackSettings
-from nodeconductor.core.models import SynchronizationStates
-from nodeconductor.core.tests.helpers import override_nodeconductor_settings
-from nodeconductor.structure import SupportedServices, signals
-from nodeconductor.structure.models import Customer, CustomerRole, VirtualMachineMixin, ProjectRole
+from nodeconductor.structure import signals
+from nodeconductor.structure.models import Customer, CustomerRole, ProjectRole
 from nodeconductor.structure.tests import factories
 
 
@@ -38,76 +33,6 @@ class SuspendServiceTest(test.APITransactionTestCase):
 
             self.assertEqual(customer.balance, amount)
 
-    # XXX: This test is too complex and tries to cover too many applications in one. It has to be  rewritten.
-    # Possible solutions:
-    #  1. Make this text abstract and override it in other applications.
-    #  2. Register factories and other test related stuff for each application and use them in this test.
-    @override_nodeconductor_settings(SUSPEND_UNPAID_CUSTOMERS=True)
-    def test_modify_suspended_services_and_resources(self):
-        self.client.force_authenticate(user=self.user)
-
-        for service_type, models in SupportedServices.get_service_models().items():
-            # XXX: quick fix for iaas cloud. Can be removed after iaas application refactoring.
-            if service_type == -1:
-                continue
-            settings = factories.ServiceSettingsFactory(
-                customer=self.customer, type=service_type, shared=True)
-
-            class ServiceFactory(factory.DjangoModelFactory):
-                class Meta(object):
-                    model = models['service']
-
-            class ServiceProjectLinkFactory(factory.DjangoModelFactory):
-                class Meta(object):
-                    model = models['service_project_link']
-
-            if service_type == SupportedServices.Types.IaaS:
-                service = ServiceFactory(customer=self.customer, state=SynchronizationStates.IN_SYNC)
-                OpenStackSettings.objects.get_or_create(
-                    auth_url='http://example.com:5000/v2',
-                    defaults={
-                        'username': 'admin',
-                        'password': 'password',
-                        'tenant_name': 'admin',
-                    })
-
-            else:
-                service = models['service'].objects.create(
-                    customer=self.customer, settings=settings, name=settings.name, available_for_all=True)
-
-            service_url = self._get_url(
-                SupportedServices.get_detail_view_for_model(models['service']), uuid=service.uuid.hex)
-
-            response = self.client.patch(service_url, {'name': 'new name'})
-            self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-            response = self.client.delete(service_url)
-            self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-            for resource_model in models['resources']:
-                if service_type == SupportedServices.Types.IaaS:
-                    continue
-
-                class ResourceFactory(factory.DjangoModelFactory):
-                    class Meta(object):
-                        model = resource_model
-
-                project = factories.ProjectFactory(customer=self.customer)
-                spl = models['service_project_link'].objects.get(service=service, project=project)
-
-                # XXX: Some resources can have more required fields and creation will fail. Lets just skip them.
-                try:
-                    resource = ResourceFactory(service_project_link=spl)
-                except:
-                    pass
-                else:
-                    resource_url = self._get_url(
-                        SupportedServices.get_detail_view_for_model(resource_model), uuid=resource.uuid.hex)
-
-                    if isinstance(resource, VirtualMachineMixin):
-                        response = self.client.post(resource_url + 'start/')
-                        self.assertEqual(response.status_code, status.HTTP_409_CONFLICT)
-
 
 class ServiceResourcesCounterTest(test.APITransactionTestCase):
     """
@@ -116,27 +41,23 @@ class ServiceResourcesCounterTest(test.APITransactionTestCase):
     For staff user resource counter should equal 2.
     """
     def setUp(self):
-        from nodeconductor.openstack.tests.factories import InstanceFactory
-        from nodeconductor.openstack.tests.factories import OpenStackServiceFactory
-        from nodeconductor.openstack.tests.factories import OpenStackServiceProjectLinkFactory
-
         self.customer = factories.CustomerFactory()
-        self.settings = factories.ServiceSettingsFactory(type=SupportedServices.Types.OpenStack, shared=True)
-        self.service = OpenStackServiceFactory(customer=self.customer, settings=self.settings)
+        self.settings = factories.ServiceSettingsFactory(shared=True)
+        self.service = factories.TestServiceFactory(customer=self.customer, settings=self.settings)
 
         self.user1 = factories.UserFactory()
         self.project1 = factories.ProjectFactory(customer=self.customer)
         self.project1.add_user(self.user1, ProjectRole.ADMINISTRATOR)
-        self.spl1 = OpenStackServiceProjectLinkFactory(service=self.service, project=self.project1)
-        self.vm1 = InstanceFactory(service_project_link=self.spl1)
+        self.spl1 = factories.TestServiceProjectLinkFactory(service=self.service, project=self.project1)
+        self.vm1 = factories.TestInstanceFactory(service_project_link=self.spl1)
 
         self.user2 = factories.UserFactory()
         self.project2 = factories.ProjectFactory(customer=self.customer)
         self.project2.add_user(self.user2, ProjectRole.ADMINISTRATOR)
-        self.spl2 = OpenStackServiceProjectLinkFactory(service=self.service, project=self.project2)
-        self.vm2 = InstanceFactory(service_project_link=self.spl2)
+        self.spl2 = factories.TestServiceProjectLinkFactory(service=self.service, project=self.project2)
+        self.vm2 = factories.TestInstanceFactory(service_project_link=self.spl2)
 
-        self.service_url = OpenStackServiceFactory.get_url(self.service)
+        self.service_url = factories.TestServiceFactory.get_url(self.service)
 
     def test_counters_for_shared_providers_should_be_filtered_by_user(self):
         self.client.force_authenticate(self.user1)