+++ nodeconductor/structure/tests/views.py
@@ -1,0 +1,18 @@
+from nodeconductor.structure import views as structure_views
+
+from . import models, serializers
+
+
+class TestServiceViewSet(structure_views.BaseServiceViewSet):
+    queryset = models.TestService.objects.all()
+    serializer_class = serializers.ServiceSerializer
+
+
+class TestServiceProjectLinkViewSet(structure_views.BaseServiceProjectLinkViewSet):
+    queryset = models.TestServiceProjectLink.objects.all()
+    serializer_class = serializers.ServiceProjectLinkSerializer
+
+
+class TestInstanceViewSet(structure_views.BaseResourceViewSet):
+    queryset = models.TestInstance.objects.all()
+    serializer_class = serializers.InstanceSerializer